Move globalEventProps to separate file from globalProps
